Parameterize the color gradient of the distance cloud according to the zenith angle

+++ Assets/Resources/XClouds/ProceduralCloud.shader
@@ -348,12 +348,12 @@ Shader "Skybox/ProceduralCloud" {
 					}
 
 					// Calculates the sun shape
-					half calcSunAttenuation(float3 lightPos, float3 ray)
+					float calcSunAttenuation(float3 lightPos, float3 ray)
 					{
 					#if SKYBOX_SUNDISK == SKYBOX_SUNDISK_SIMPLE
-						half3 delta = lightPos - ray;
-						half dist = length(delta);
-						half spot = 1.0 - smoothstep(0.0, _SunSize, dist);
+						float3 delta = lightPos - ray;
+						float dist = length(delta);
+						float spot = 1.0 - smoothstep(0.0, _SunSize, dist);
 						return spot * spot;
 					#else // SKYBOX_SUNDISK_HQ
 						float focusedEyeCos = pow(saturate(dot(lightPos, ray)), _SunSizeConvergence);
